Even more warnings gone

+++ src-core/modules/meteor/deframer.cpp
@@ -54,7 +54,7 @@ namespace meteor
         return state;
     }
 
-    std::vector<std::array<uint8_t, CADU_SIZE>> CADUDeframer::work(uint8_t *input, size_t size)
+    std::vector<std::array<uint8_t, CADU_SIZE>> CADUDeframer::work(uint8_t *input, int size)
     {
         // Output buffer
         std::vector<std::array<uint8_t, CADU_SIZE>> frames;
